LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/svg-graphics-element-003-expected.txt


FAIL View timeline attached to SVG graphics element assert_equals: expected "rgb(0, 64, 127)" but got "rgb(0, 128, 0)"
